+++ test/__unit__.js
@@ -4,4 +4,6 @@ const MockBrowser = require('mock-browser')
 global.XMLHttpRequest = XMLHttpRequest.XMLHttpRequest
 global.window = MockBrowser.mocks.MockBrowser.createWindow()
 
-global.location = 'http://127.0.0.1:8988'
+global.location = {
+  origin: 'http://127.0.0.1:8988',
+}